Move setGutterMarker, clearGutter and lineInfo to Doc
