+++ src/edit/methods.js
@@ -1,5 +1,4 @@
 import { deleteNearSelection } from "./deleteNearSelection"
-import { changeLine } from "../model/changes"
 import { commands } from "./commands"
 import { attachDoc } from "../model/document_data"
 import { activeElt, addClass, rmClass } from "../util/dom"
@@ -18,9 +17,9 @@ import { addToScrollPos, calculateScrollPos, ensureCursorVisible, resolveScrollT
 import { heightAtLine } from "../line/spans"
 import { updateGutterSpace } from "../display/update_display"
 import { lineLeft, lineRight, moveLogically, moveVisually } from "../util/bidi"
-import { indexOf, insertSorted, isEmpty, isWordChar, sel_dontScroll, sel_move } from "../util/misc"
+import { indexOf, insertSorted, isWordChar, sel_dontScroll, sel_move } from "../util/misc"
 import { signalLater } from "../util/operation_group"
-import { getLine, isLine, lineAtHeight, lineNo } from "../line/utils_line"
+import { getLine, isLine, lineAtHeight } from "../line/utils_line"
 import { regChange, regLineChange } from "../display/view_tracking"
 
 // The publicly visible API. Note that methodOp(f) means
@@ -217,43 +216,6 @@ export default function(CodeMirror) {
 
     defaultTextHeight: function() { return textHeight(this.display) },
     defaultCharWidth: function() { return charWidth(this.display) },
-
-    setGutterMarker: methodOp(function(line, gutterID, value) {
-      return changeLine(this.doc, line, "gutter", line => {
-        let markers = line.gutterMarkers || (line.gutterMarkers = {})
-        markers[gutterID] = value
-        if (!value && isEmpty(markers)) line.gutterMarkers = null
-        return true
-      })
-    }),
-
-    clearGutter: methodOp(function(gutterID) {
-      let doc = this.doc, i = doc.first
-      doc.iter(line => {
-        if (line.gutterMarkers && line.gutterMarkers[gutterID]) {
-          line.gutterMarkers[gutterID] = null
-          regLineChange(this, i, "gutter")
-          if (isEmpty(line.gutterMarkers)) line.gutterMarkers = null
-        }
-        ++i
-      })
-    }),
-
-    lineInfo: function(line) {
-      let n
-      if (typeof line == "number") {
-        if (!isLine(this.doc, line)) return null
-        n = line
-        line = getLine(this.doc, line)
-        if (!line) return null
-      } else {
-        n = lineNo(line)
-        if (n == null) return null
-      }
-      return {line: n, handle: line, text: line.text, gutterMarkers: line.gutterMarkers,
-              textClass: line.textClass, bgClass: line.bgClass, wrapClass: line.wrapClass,
-              widgets: line.widgets}
-    },
 
     getViewport: function() { return {from: this.display.viewFrom, to: this.display.viewTo}},
